added authentification and authorization

+++ frontend/src/components/SportDetails.jsx
@@ -1,21 +1,32 @@
 import moment from "moment"
 import { useSportHook } from "./context/SportContext"
 import {FaRegTrashAlt, FiEdit} from "react-icons/all"
+import { useAuthContext } from "./context/AuthContext"
 
 
 
-const SportDetails = ({title, reps, load, createdAt, _id,}) => {
-  const {dispatch, fetchSports, sports, 
+const SportDetails = ({title, reps, load, createdAt, _id, fetchSports}) => {
+  const {dispatch, 
     setTitle,
     setReps,
     setLoad,
   } = useSportHook()
+
+const {user} = useAuthContext()
+
   // console.log(sports)
 
   const handleErase = async ()=>{
 
+    if(!user){
+      return
+    }
+
     const response = await fetch(`http://localhost:8000/api/sports/${_id}`,{
-      method: "DELETE"
+      method: "DELETE",
+      headers:{
+      "Authorization": `Bearing ${user.token}`
+      }
     })
     const data = response.json();
     console.log(data);
@@ -25,14 +36,15 @@ const SportDetails = ({title, reps, load, createdAt, _id,}) => {
     fetchSports()
   }
 
-  const handleEdit = async()=>{
+  const handleEdit = async(e)=>{
     
     const response = await fetch(`http://localhost:8000/api/sports/${_id}`,{
       method: "PATCH",
       body: JSON.stringify({ title, reps, load }),
       headers: {
         'Accept': 'application/json',
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+      "Authorization": `Bearing ${user.token}`
     },
     })
     const data = await response.json()
@@ -41,7 +53,6 @@ const SportDetails = ({title, reps, load, createdAt, _id,}) => {
       setTitle(data.title)
       setReps(data.reps)
       setLoad(data.load)
-      dispatch({type:"UPDATE", payload:data})
     }
   }
